drivers: mfd: s2mpg13: mask ibi regions at init
Mask all ibi regions during init.
Unmask PM and METER region interrupts only when registered.
Handle error conditions.
Set interrupt masks before registering IRQs with genirq.

Bug: 214317583
Test: build and boot on device. Check s2mpg12/13 interrupts
under '/proc/interrupts'

+++ include/linux/mfd/samsung/s2mpg13.h
@@ -118,8 +118,8 @@ struct s2mpg13_dev {
 
 	/* mutex for s2mpg13 irq handling */
 	struct mutex irqlock;
-	int irq_masks_cur[S2MPG13_IRQ_GROUP_NR];
-	int irq_masks_cache[S2MPG13_IRQ_GROUP_NR];
+	u8 irq_masks_cur[S2MPG13_IRQ_GROUP_NR];
+	u8 irq_masks_cache[S2MPG13_IRQ_GROUP_NR];
 
 	/* Work queue */
 	struct workqueue_struct *irq_wqueue;
